Rename 'embedchain/vectordb' to 'embedchain/vector_db'
