+++ embedchain/config/apps/custom_app_config.py
@@ -3,7 +3,7 @@ from typing import Optional
 from dotenv import load_dotenv
 
 from embedchain.helper.json_serializable import register_deserializable
-from embedchain.vectordb.base import BaseVectorDB
+from embedchain.vector_db.base import BaseVectorDB
 
 from .base_app_config import BaseAppConfig
